requirements.txt
Django==2.2.16
djangorestframework
pytz==2018.5
django-rest-passwordreset
